Refactor: Define distinct build functions for 'matrixmath' and 'matrixmath_serializer' projects.
