+++ .github/build.sh
@@ -10,64 +10,64 @@
 # @usage
 # Run this script from the root of your project:
 #   ./build.sh;
-#
-# Global settings:
-# - BASE_NAME: Base name for the project.
-# - PROJECT_PATH: Root path of the project.
-# - TEST_DEPENDENCIES: Dependencies for tests.
-# - LIBRARY_DEPENDENCIES: Dependencies for libraries.
-# - LIBRARY_CODE_SEARCH_PATHS: Search paths for library code.
-# - TEST_CODE_SEARCH_PATHS: Search paths for test code.
-# - BASE_BUILD_PATH: Base build directory.
-# - LIBRARY_BUILD_PATH: Build directory for libraries.
-# - TEST_BUILD_PATH: Build directory for tests.
-# - BIN_PATH: Output directory for binaries.
-# - APP_NAME: Name of the main application.
 
-# Determine the directory of the script
+
+# Determine the directory of the script.
 SCRIPT_DIR=$(dirname "$(readlink -f "$0")");
 # Load helper functions
 source "$SCRIPT_DIR/helper.sh";
 
-# Global Settings.
-BASE_NAME='libmatrixmath';   # Base name for the project.
-PROJECT_PATH=$(pwd);         # Root path of the project.
-
-# Dependencies for tests and library (add as needed).
-TEST_DEPENDENCIES='-lm -ljson -lstr';
-LIBRARY_DEPENDENCIES='-lm -ljson -lstr';
-
-# Search paths for library and test code.
-LIBRARY_CODE_SEARCH_PATHS="$PROJECT_PATH/include $PROJECT_PATH/src";
-TEST_CODE_SEARCH_PATHS="$LIBRARY_CODE_SEARCH_PATHS $PROJECT_PATH/tests";
-
-# Build paths.
-BASE_BUILD_PATH="$PROJECT_PATH/build";
-LIBRARY_BUILD_PATH="$BASE_BUILD_PATH/library";
-TEST_BUILD_PATH="$BASE_BUILD_PATH/test";
-
-# Output paths.
+# Specify the root path of the project.
+PROJECT_PATH=$(pwd);
 BIN_PATH="$PROJECT_PATH/bin";
-APP_NAME="$BASE_NAME.app";
 
-# Ensure that library dependencies are installed on the local system.
-if ! [ -f /usr/local/lib/libstr.so ] || ! [ -f /usr/local/lib/libjson.so ]; then
-  sudo "$SCRIPT_DIR/install_from_remote.sh";
-fi
+# Build 'matrixmath' project.
+#
+# This function builds the 'matrixmath' project, including the main application and libraries.
+#
+# Arguments:
+#   $1 - Root path of the project.
+#
+# Usage:
+#   build_matrixmath_project "/path/to/project";
+build_matrixmath_project() {
+  # Project Settings.
+  local project_path="$1"; # Root path of the project.
+  local base_name='libmatrixmath';   # Base name for the project.
+  local test_dependencies='-lm'; # Dependencies for tests (add as needed).
+  local library_dependencies='-lm'; # Dependencies for library (add as needed).
+  local namespace='matrixmath'; # The project namespace.
+
+  # Build the project.
+  build_project "$base_name" "$project_path" "$test_dependencies" "$library_dependencies" "$namespace";
+}
+
+# Build 'matrixmath_serializer' project.
+#
+# This function builds the 'matrixmath_serializer' project, including the main application and libraries.
+#
+# Arguments:
+#   $1 - Root path of the project.
+#
+# Usage:
+#   build_matrixmath_serializer_project "/path/to/project";
+build_matrixmath_serializer_project() {
+  # Project Settings.
+  local project_path="$1"; # Root path of the project.
+  local base_name='libmatrixmath_serializer';   # Base name for the project.
+  local test_dependencies='-lmatrixmath -ljson -lstr'; # Dependencies for tests (add as needed).
+  local library_dependencies='-lmatrixmath -ljson -lstr'; # Dependencies for library (add as needed).
+  local namespace='serialization'; # The project namespace.
+
+  # Build the project.
+  build_project "$base_name" "$project_path" "$test_dependencies" "$library_dependencies" "$namespace";
+}
 
 # Clean up build directory before start the build.
-clean_directory "$BASE_BUILD_PATH";
-
-# Build the main app.
-app_files_to_compile=$(get_files_to_compile "$TEST_CODE_SEARCH_PATHS");
-build_app "$app_files_to_compile" $TEST_BUILD_PATH $APP_NAME "$TEST_DEPENDENCIES" $BIN_PATH;
-
-# Create shared and static libraries.
-library_files_to_compile=$(get_files_to_compile "$LIBRARY_CODE_SEARCH_PATHS");
-create_libraries "$library_files_to_compile" $LIBRARY_BUILD_PATH $BASE_NAME "$LIBRARY_DEPENDENCIES" $BIN_PATH;
-
-# Clean precompiled header files from the project directories.
-clean_project_precompiled_headers "$PROJECT_PATH" > /dev/null;
-
-# Test the APP execution.
-"$BIN_PATH/$APP_NAME";
+clean_directory "$BIN_PATH";
+# Build 'matrixmath' project.
+build_matrixmath_project "$PROJECT_PATH";
+# Install the project in the local system from source.
+sudo "$SCRIPT_DIR/install_from_local.sh" >> /dev/null;
+# Build 'matrixmath_serializer' project.
+build_matrixmath_serializer_project "$PROJECT_PATH";
