+++ src/serialization/matrix_serializer.c
