+++ src/serialization/vector_serializer.c
